Collapse the database access protocols into DatabaseReader and DatabaseWriter (#935)
The database access layer was split into six per-capability protocols
(RecordReader, RecordLocator, MetricReader, ActivityReader,
RetentionReader, RecordWriter) glued together by typealiases, yet no
consumer ever used the narrow protocols as types — providers take
DatabaseReader and sync takes any Database. This collapses them into two
real protocols, DatabaseReader and DatabaseWriter, keeping
readAll/readMore/metricSeries as extension conveniences and both read
overloads as requirements since the backends pick different default
limits. The activity and retention requirements move out of the UI
provider files into Core, the seven tiny Hosted/Access conformance files
merge into HostedReader/HostedWriter, and SeriesScalar folds into
MetricScalar, which only Int and Double conform to. Behavior is
unchanged; stacked on #933.
