+++ Sources/Scout/Core/Database/Access/Database.swift
@@ -9,5 +9,48 @@ import Foundation
 
 typealias Database = DatabaseReader & DatabaseWriter
 
-typealias DatabaseReader = RecordLocator & MetricReader & ActivityReader & RetentionReader & Sendable
-typealias DatabaseWriter = RecordWriter & Sendable
+protocol DatabaseReader: Sendable {
+    func read(matching query: RecordQuery, fields: [String]?) async throws -> RecordChunk
+    func read(matching query: RecordQuery, fields: [String]?, limit: Int) async throws -> RecordChunk
+    func lookup(recordName: String, fields: [String]?) async throws -> Record
+    func series(matching query: SeriesQuery) async throws -> [MetricSeries]
+    func activity(in range: Range<Date>) async throws -> [ActivityPoint]
+    func retention(in range: Range<Date>) async throws -> [RetentionCohort]
+}
+
+protocol DatabaseWriter: Sendable {
+    func write(record: Record) async throws
+    func write(records: [Record]) async throws
+}
+
+extension DatabaseReader {
+    func read(matching query: RecordQuery, fields: [String]?, limit: Int) async throws -> RecordChunk {
+        try await read(matching: query, fields: fields)
+    }
+
+    func readMore(from cursor: RecordCursor, fields: [String]?) async throws -> RecordChunk {
+        try await cursor.next(fields)
+    }
+
+    func readAll(matching query: RecordQuery, fields: [String]?) async throws -> [Record] {
+        var chunk = try await read(matching: query, fields: fields)
+        while let cursor = chunk.cursor {
+            chunk += try await readMore(from: cursor, fields: fields)
+        }
+        return chunk.records
+    }
+
+    func readAll<T: RecordDecodable>(matching query: RecordQuery, fields: [String]? = nil) async throws -> [T] {
+        try await readAll(matching: query, fields: fields).map(T.init)
+    }
+}
+
+extension DatabaseWriter {
+    // Records are written in batches no larger than this; the backends cap a single
+    // save/modify request at 400 records.
+    static var maxBatchSize: Int { 400 }
+}
+
+struct RecordNotFoundError: LocalizedError {
+    let errorDescription: String? = "No record found for the requested identifier"
+}
