+++ Sources/Scout/Core/Diagnostics/Metrics/MetricScalar.swift
@@ -9,12 +9,16 @@ import CoreData
 
 protocol MetricScalar: AdditiveArithmetic & Comparable & Hashable & Sendable & RecordValueConvertible {
     associatedtype Object: MetricsValued where Object.Value == Self
+    static var seriesValues: String { get }
+    init(_ value: Double)
 }
 
 extension Int: MetricScalar {
     typealias Object = IntMetricsEntry
+    static var seriesValues: String { "int" }
 }
 
 extension Double: MetricScalar {
     typealias Object = DoubleMetricsEntry
+    static var seriesValues: String { "double" }
 }
