+++ Sources/Scout/UI/Analytics/Activity/ActivityProvider.swift
@@ -15,10 +15,6 @@ class ActivityProvider: ObservableObject, Provider {
     }
 }
 
-protocol ActivityReader: RecordReader {
-    func activity(in range: Range<Date>) async throws -> [ActivityPoint]
-}
-
 struct ActivityPoint: Decodable {
     let date: Int64
     let dau: Int
